Move project Interface package to be easily access by classes
