+++ src/main/java/controllers/Interfaces/ADSInterface.java
@@ -1,7 +1,6 @@
-package ADS;
+package Interfaces;//package ADS;
 
 import com.pi4j.io.gpio.event.GpioPinAnalogValueChangeEvent;
-import com.pi4j.io.gpio.event.GpioPinListener;
 
 public interface ADSInterface<T> {
     void setupGpio();
