+++ src/main/java/controllers/Interfaces/setValueInterface.java
@@ -1,5 +1,4 @@
-package ServoBlaster;
-
+package Interfaces;
 public interface setValueInterface {
     void setSliderValue(int T) throws InterruptedException;
 }
