+++ src/main/java/main.iml
@@ -4,7 +4,7 @@
     <exclude-output />
     <content url="file://$MODULE_DIR$">
       <sourceFolder url="file://$MODULE_DIR$/controllers" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/interfaces" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/views" isTestSource="false" />
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
